requirements.txt
asgiref==3.8.1
cffi==1.16.0
cryptography==43.0.0
Django==5.0.7
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
pycparser==2.22
PyJWT==2.8.0
sqlparse==0.5.1
tzdata==2024.1
